Revert name change to Snowball back to IceBall
Changing the name for a project is a difficult thing to do and creates multiple challenges, such as in this instance where we imported @woodrow73's new code for this plugin into our existing codebase for IceBall.

When we imported the new code in, there were several issues with the plugin not realizing the name of the packages, where all of the classes were, and what variables could be found in what classes. The "easy" fix was to revert the name of the plugin back to IceBall to ensure that the project is maintainable and will compile successfully in the future.

There is a still a lot of cleaning up that needs to be done in the main class, especially in terms of file management and storing files in the appropriate server directories.